+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## What's new in 3.8.9
 
+### Features
+
+- Added the possibility to remove a reviewer from a PR
+
 ### Bug Fixes
 
 - Removed setting options that had no practical or business logic use.